virglrenderer: add virgl_renderer_resource_type to virgl_renderer_resource_info_ext
In order to support other native texture types, we need to extend the
struct. We use an anonymous union type in order for the memory layout to
match prior versions. However, we bump
VIRGL_RENDERER_RESOURCE_INFO_EXT_VERSION in order to support the additional
member `native_type`.

Using v1 headers with v0 shared library: the client must check the
`version` field and look at `native_type` only if it is >= 1. Otherwise,
it can only safely assume `d3d_tex2d` is either NULL or non-NULL.

Using v1 headers with an outdated client: because of the anonymous union,
it is still valid to refer to `info->d3d_tex2d` and it will not break the
compilation even without client updates.

Using v0 headers with an updated client: the memory layout of the struct is
still the same so `info->d3d_tex2d` should still work.