proxy: introduce in-process rendering
iOS does not allow for fork() or use of shmem so we will need to keep
rendering in the same process. When 'thread' render server worker is
selected, we will now support init with !VIRGL_RENDERER_RENDER_SERVER
by creating a new thread in the host process to act as the server.